+++ src/monero/subscriptions.rs
@@ -174,7 +174,6 @@ pub async fn check_monero_subscriptions(
                 update_subscription(
                     db_client,
                     subscription.id,
-                    subscription.chain_id.inner(),
                     &expires_at,
                     &Utc::now(),
                 ).await?;